Move more files.
