+++ nbodykit/algorithms/convpower/__init__.py
@@ -1,0 +1,4 @@
+
+from .fkp import ConvolvedFFTPower
+from .catalog import FKPCatalog
+
