+++ nbodykit/algorithms/convpower/fkp.py
@@ -749,8 +749,9 @@ def _cast_mesh(mesh, Nmesh):
     """
     Cast an object to a MeshSource. Nmesh is used only on FKPCatalog
     """
-    from nbodykit.source.catalog import FKPCatalog
-    from nbodykit.source.catalogmesh import FKPCatalogMesh
+    from .catalog import FKPCatalog
+    from .catalogmesh import FKPCatalogMesh
+
     if not isinstance(mesh, (FKPCatalogMesh, FKPCatalog)):
         raise TypeError("input sources should be a FKPCatalog or FKPCatalogMesh")
